# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/login": {
            "post": {
                "tags": [
                    "Todo List"
                ],
                "summary": "login",
                "description": "Returns token",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "Todo List"
                ],
                "summary": "register",
                "description": "Returns token",
                "operationId": "register",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable"
                    }
                }
            }
        },
        "/tasks": {
            "get": {
                "tags": [
                    "Todo List"
                ],
                "summary": "Get list of tasks",
                "description": "Returns list of tasks",
                "operationId": "getTasksList",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Example: ?search=John",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "Example: ?filter=id;title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "orderBy",
                        "in": "query",
                        "description": "Example: ?orderBy=id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sortedBy",
                        "in": "query",
                        "description": "Example: ?sortedBy=desc",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TaskTransformer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Todo List"
                ],
                "summary": "Store new task",
                "description": "Returns task data",
                "operationId": "storeTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable"
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "tags": [
                    "Todo List"
                ],
                "summary": "Get task information",
                "description": "Returns task data",
                "operationId": "getTaskById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Todo List"
                ],
                "summary": "Update existing task",
                "description": "Returns updated task data",
                "operationId": "updateTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateTaskRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/task/{id}": {
            "delete": {
                "tags": [
                    "Todo List"
                ],
                "summary": "Delete existing task",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteTask",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/tasks/{id}/change-to-done": {
            "patch": {
                "tags": [
                    "Todo List"
                ],
                "summary": "Change task status",
                "description": "Returns updated task data",
                "operationId": "changeTaskStatusToDone",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Task"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateTaskRequest": {
                "title": "Create task request",
                "description": "Create task request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "title": {
                        "title": "Title",
                        "description": "Title of the new task",
                        "type": "string",
                        "example": "A nice task"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the new task",
                        "type": "string",
                        "example": "This is new task's description"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "title": "Login request",
                "description": "Login request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "email": {
                        "title": "Email",
                        "description": "Email",
                        "type": "string",
                        "example": "test@test.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Task": {
                "title": "Task",
                "description": "Task model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "title": {
                        "title": "Title",
                        "description": "Title of the new task",
                        "type": "string",
                        "example": "A nice task"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the new task",
                        "type": "string",
                        "example": "This is new task's description"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-01-27 17:50:45"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Task"
                }
            },
            "User": {
                "title": "User",
                "description": "Authorization token",
                "properties": {
                    "token": {
                        "title": "Token",
                        "description": "Authorization token",
                        "type": "string",
                        "example": "3|CWCXh9LqmZp3XJSIRqxU0hirKTrhvr2rqjEos5Mw"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
            "RegisterRequest": {
                "title": "Register request",
                "description": "Register request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email",
                        "type": "string",
                        "example": "test@test.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "TaskTransformer": {
                "title": "TaskTransformer",
                "description": "Task resource",
                "properties": {
                    "data": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Task"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TaskTransformer"
                }
            },
            "UpdateTaskRequest": {
                "title": "Update task request",
                "description": "Update task request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "title": {
                        "title": "Title",
                        "description": "Title of the new task",
                        "type": "string",
                        "example": "A nice task"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description of the new task",
                        "type": "string",
                        "example": "This is new task's description"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Todo List",
            "description": "API Endpoints of Projects"
        }
    ]
}